image controller + image service

+++ src/main/java/com/shareanycar/service/ImageService.java
@@ -4,6 +4,7 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.util.List;
 
 import javax.inject.Inject;
 
@@ -18,41 +19,40 @@ import com.shareanycar.util.MiscUtils;
 
 @Service
 public class ImageService {
-	
+
 	@Inject
 	public CarDao carDao;
-	
+
 	@Inject
 	public ImageDao imageDao;
-	
+
 	@Inject
 	public AppConfig appConfig;
 
 	@Inject
 	public MiscUtils miscUtils;
-	
-	private String saveFile(InputStream is) throws Exception{
-		
-		String uploadedFileLocation =  appConfig.getImageLocation();
+
+	private String saveFile(InputStream is) throws Exception {
+
+		String uploadedFileLocation = appConfig.getImageLocation();
 		String fileName = miscUtils.randonString();
-				
+
 		OutputStream out = new FileOutputStream(new File(uploadedFileLocation + fileName));
-					
+
 		int read = 0;
 		byte[] bytes = new byte[1024];
-		
-				
+
 		while ((read = is.read(bytes)) != -1) {
-				out.write(bytes, 0, read);
+			out.write(bytes, 0, read);
 		}
-			
+
 		out.flush();
 		out.close();
-		
+
 		return fileName;
 	}
 
-	public Image uploadImage(Long carId, InputStream is) throws Exception {
+	public Image uploadImage(Long carId, Long ownerId, InputStream is) throws Exception {
 
 		Car car = carDao.findOne(carId);
 
@@ -60,31 +60,44 @@ public class ImageService {
 			throw new Exception("can not find car with id:" + carId);
 		}
 
+		if (car.getOwner().getId() != ownerId) {
+			throw new Exception(
+					"can not load image, car does not belong to current owner id:" + ownerId + "; car id" + carId);
+		}
+
 		String fileName = saveFile(is);
-		
+
 		Image image = new Image();
-		
+
 		image.setName(fileName);
 		image.setUrl(appConfig.getUrlPrefix() + fileName);
 		image.setCar(car);
-		
+
 		image = imageDao.save(image);
 		return image;
 	}
-	
-	public boolean deleteImage(Long id){
-		Image image = imageDao.findOne(id);
-		if(image == null){
-			return false;
+
+	public boolean deleteImage(Long imageId, Long ownerId) throws Exception {
+		Image image = imageDao.findOne(imageId);
+		if (image == null) {
+			throw new Exception("can not find image with id:" + imageId);
 		}
-		
+
+		if (image.getCar().getOwner().getId() != ownerId) {
+			throw new Exception("image belongs to car that does not belong to current owner id:" + ownerId
+					+ "; image id" + imageId);
+		}
+
 		File f = new File(appConfig.getImageLocation() + image.getName());
-		if(f.delete()){
+		if (f.delete()) {
 			imageDao.delete(image);
 			return true;
-		}else{
+		} else {
 			return false;
 		}
-		
+	}
+	
+	public List<Image> findImageByCarId(Long carId){
+		return imageDao.findImageByCarId(carId);
 	}
 }
